move car_models to utilties
