+++ backend/src/cleaners/cleaner.py
@@ -1,4 +1,4 @@
-from ..utils import database as db
+from ..utilities import database as db
 from . import facebook as fb
 from . import craigslist as cl
 from . import utils
