+++ backend/src/cleaners/utils.py
@@ -2,6 +2,7 @@ import re
 import json
 from difflib import SequenceMatcher, get_close_matches
 
+
 def clean_currency(price_str):
     clean_str = price_str.replace("$", "").replace(",", "")
 
@@ -79,44 +80,48 @@ def best_fitting_make(title):
     best_match = None
 
     for make in CAR_MAKES:
-        similarity = max(SequenceMatcher(None, make, word).ratio() for word in title_words)
+        similarity = max(
+            SequenceMatcher(None, make, word).ratio() for word in title_words
+        )
         if similarity > max_similarity:
             max_similarity = similarity
             best_match = make
 
     return best_match
 
-# TODO: Call this from craigslist and facebook
+
 def extract_model(title, make, regex):
-    with open("car_models.json") as models_json:
+    with open("../utilities/car_models.json") as models_json:
         models = json.load(models_json)
         models = models[make]
 
     match = re.search(regex, title)
-    
+
     if match:
         model_search_area = match.group(1)
         model_search_area = model_search_area.replace("-", "")
-    
+
         close_match = get_close_matches(model_search_area, models, n=1, cutoff=0.4)
         if close_match:
             return close_match[0]
-       
+
     return incremental_model_search(title, make, models)
 
 
 def incremental_model_search(title, make, models):
     title_words = title.split(" ")
     search_substring = []
-    
+
     for word in title_words:
         if word.lower() in make.lower():
             continue
         search_substring.append(word)
-        
-        close_match = get_close_matches(" ".join(search_substring), models, n=1, cutoff=0.4)
-        
+
+        close_match = get_close_matches(
+            " ".join(search_substring), models, n=1, cutoff=0.4
+        )
+
         if close_match:
             return close_match[0]
-    
+
     return None
